requirements.txt
absl-py==1.4.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
biopython==1.81
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
cmake==3.26.1
comm==0.1.3
contextlib2==21.6.0
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
dm-haiku==0.0.9
dm-tree==0.1.8
executing==1.2.0
filelock==3.11.0
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
jax==0.4.8
jaxlib==0.4.7+cuda12.cudnn88
jedi==0.18.2
Jinja2==3.1.2
jmp==0.0.4
jupyter_client==8.1.0
jupyter_core==5.3.0
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
lit==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
ml-collections==0.1.1
ml-dtypes==0.0.4
mock==5.0.1
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.0.26
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvcc-cu12==12.1.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.55
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.8.1.3
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.4
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.4.55
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.0.2.55
nvidia-nccl-cu11==2.14.3
nvidia-nvjitlink-cu12==12.1.55
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.0
parso==0.8.3
pdb-tools==2.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
pyrosetta==2023.14+release.7132bdc
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
torch==2.0.0
tornado==6.2
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.14.1
